Update autoported files. Fix manually ported files: TableRendrer and TypogrphyUtils

+++ itext/itext.io/itext/io/codec/TIFFDirectory.cs
@@ -379,7 +379,7 @@ namespace iText.IO.Codec {
                 return null;
             }
             else {
-                return fields[i];
+                return fields[(int)i];
             }
         }
